Refactor to move bulk of reducer logic to helpers
This is necessary because we are going to change a lot of logic under
the hood and we need to carefully manage those changes.
