+++ .watchmanconfig
@@ -1,1 +1,3 @@
-{}
+{
+    "ignore_dirs":[".git"]
+}
